requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.5
pkg-resources==0.0.0
python-gitlab==0.21.2
requests==2.18.3
six==1.10.0
urllib3==1.22
